+++ kernel/src/arch/x86_64/mod.rs
@@ -38,7 +38,3 @@ pub use crate::arch::elf::X86_64ElfArch as CurrentElfArch;
 pub use crate::arch::smp::X86_64SMPArch as CurrentSMPArch;
 
 pub use crate::arch::sched::X86_64SchedArch as CurrentSchedArch;
-
-pub use crate::arch::mm::fault::X86_64PageFault as PageFaultArch;
-
-pub use crate::arch::mm::pkey::X86_64ProtectionKey as ProtectionKey;
